+++ pages/others/components/Preferences.tsx
