+++ pages/others/components/Qualifications.tsx
